requirements.txt
Django==3.0.2
google-cloud-vision==1.0.0
pillow